pip is going to be removing support for dependency links.
This commit flag is necessary for compatibility with pip 1.5, but it's
unrecognized and error-throwing in earlier versions. In 1.6 it will
apparently be entirely removed.

I think the recommended solution is to not rely on this information
being available from setup.py, and just specify it in our Makefile's
invocation of pip.

The relevant documentation is under development here:
http://www.pip-installer.org/en/develop/dependency_links.html